Fix: cRef must be word-size-native UInt, not hardcoded UInt64
The component handle `cRef` is round-tripped through `Ptr{Nothing}` to
recover the Julia object from a plain integer (FMI2/struct.jl,
FMI3/struct.jl, eval.jl). It was hardcoded as UInt64, but on 32-bit
Julia a real pointer is 32 bits wide, so `Ptr{Nothing}(cRef::UInt64)`
has no matching constructor and simulation fails immediately with
`MethodError: no method matching Ptr{Nothing}(::UInt64)`.

Using UInt (word-size-native: UInt64 on 64-bit, UInt32 on 32-bit)
matches pointer width on both architectures.

This surfaced after fixing the related FMICore.jl bug where
fmi2Real/fmi3Float64 was incorrectly Cfloat on 32-bit -- once that's
fixed, this cRef bug is the next thing blocking 32-bit FMU simulation.

Confirmed fix resolves simulation failures on Julia 1.12.7 x86 for
multiple independent FMUs.

+++ src/FMI2/struct.jl
@@ -42,7 +42,7 @@ The mutable struct represents an allocated instance of an FMU in the FMI 2.0.2 S
 """
 mutable struct FMU2Component{F} <: FMUInstance
     addr::fmi2Component
-    cRef::UInt64
+    cRef::UInt
 
     fmu::F
     state::fmi2ComponentState
@@ -141,7 +141,7 @@ mutable struct FMU2Component{F} <: FMUInstance
     # constructor
     function FMU2Component{F}() where {F}
         inst = new{F}()
-        inst.cRef = UInt64(pointer_from_objref(inst))
+        inst.cRef = UInt(pointer_from_objref(inst))
         inst.state = fmi2ComponentStateInstantiated
         inst.t = NO_fmi2Real
         inst.t_offset = fmi2Real(0.0)
